+++ crmGUI/Model/BaseModel.py
