+++ crmGUI/starting.py
@@ -1,5 +1,6 @@
-from src.View.MainWindow import MainWindow
+from crmGUI.View.Screens.MainWindow import MainWindow
 from PyQt5.QtWidgets import QApplication
+
 
 class Start:
     def __init__(self):
@@ -7,3 +8,7 @@ class Start:
         window = MainWindow()
         window.show()
         app.exec_()
+
+
+if __name__ == '__main__':
+    startApp = Start()
