--- main.py
@@ -1,6 +1,0 @@
-from src.starting import Start
-
-if __name__ == '__main__':
-    startApp = Start()
-
-
